Objectified grammar adt, with equivalent output as last build.
Grammar class is extended in two ways: InputGrammar, and ParsingGrammar.
ParserApp makes an instance of each: G : ParsingGrammar, and G_prime : InputGrammar.
The precedence table is now called precedenceMatrix (protected) and has methods for getting and setting, getRel, setRel.
The constructor for ParsingGrammar fills in the precedence matrix and productions table automatically, but the InputGrammar will need to use methods to compute these things, as well as Left/Right sets.
Then we will be able to model the expectedResults with ParserApp.
ParserApp still contains the semanticStack, and semanticActions methods but ParsingGrammar contains the functions for converting to and from symbol types and names.